==> Course 4 - Pandas (Analyzing and Processing Data)/1 - Base de Dados.ipynb <==
# Relatório de Análise I

## Importando a Base de Dados

import pandas as pd

# importando
pd.read_csv('dados/aluguel.csv', sep = ';')

dados = pd.read_csv('dados/aluguel.csv', sep = ';')
dados

type(dados)

dados.info()

dados.head(10)

## Informações gerais sobre a Base de Dados

dados.dtypes

tipos_de_dados = pd.DataFrame(dados.dtypes, columns = ['Tipos de Dados'])

tipos_de_dados.columns.name = 'Variáveis'

tipos_de_dados

dados.shape #tupla --> primeiro número é a quantidade de registros, segundo quantas variáveis existem

dados.shape[0]

dados.shape[1]

print(f'A base de dados apresenta {dados.shape[0]} registros (imóveis) e {dados.shape[1]} variáveis')

import pandas as pd
data = [['Fulano', 12, 7.0, True],
        ['Sicrano', 15, 3.5, False], 
        ['Beltrano', 18, 9.3, True]]
dados = pd.DataFrame(data, 
        columns = ['Aluno', 'Idade', 'Nota', 'Aprovado'])
dados

tipos_de_dados = pd.DataFrame(dados.dtypes, 
    columns = ['Tipos de Dados'])
tipos_de_dados.columns.name = 'Variáveis'
tipos_de_dados
