==> Course 4 - Pandas (Analyzing and Processing Data)/3 - Imóveis Residenciais.ipynb <==
import pandas as pd

# Relatório de Análise III

## Imóveis Residenciais

dados = pd.read_csv('dados/aluguel.csv', sep = ';')

dados.head(10)

list(dados['Tipo'].drop_duplicates())

residencial = ['Quitinete',
 'Casa', 
 'Apartamento',
 'Casa de Condomínio',
 'Casa de Vila']

residencial

dados.head(10)

selecao =  dados['Tipo'].isin(residencial)
selecao

dados_residencial = dados[selecao]

dados_residencial

list(dados_residencial['Tipo'].drop_duplicates())

dados_residencial.shape[0]

dados.shape[0]

dados_residencial.index = range(dados_residencial.shape[0])

dados_residencial

### Exercício

numeros = [i for i in range(11)]
letras = [chr(i + 65) for i in range(11)]
nome_coluna = ['N']

df = pd.DataFrame(data = numeros, index = letras, columns = nome_coluna)
df

selecao = df['N'].isin([i for i in range(11) if i % 2 == 0])
df = df[selecao]
df

## Exportando a Base de Dados

dados_residencial

dados_residencial.to_csv('dados/aluguel_residencial.csv', sep = ';')

dados_residencial2 = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')

dados_residencial2

dados_residencial.to_csv('dados/aluguel_residencial.csv', sep = ';', index = False)

dados_residencial2 = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')

dados_residencial2
