==> Course 4 - Pandas (Analyzing and Processing Data)/4 - Seleções e Frequências.ipynb <==
# Relatório de Análise IV

## Seleções e Frequências

import pandas as pd

dados = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')

dados.head(10)

# Selecione somente os imóveis classificados com tipo 'Apartamento'.
selecao = dados['Tipo'] == 'Apartamento'
n1 = dados[selecao].shape[0]
n1

# Selecione os imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila'.
selecao = (dados['Tipo'] == 'Casa') | (dados['Tipo'] == 'Casa de Condomínio') | (dados['Tipo'] == 'Casa de Vila')
n2 = dados[selecao].shape[0]
n2

# Selecione os imóveis com área entre 60 e 100 metros quadrados, incluindo os limites.
selecao = (dados['Area'] >= 60) & (dados['Area'] <= 100)
n3 = dados[selecao].shape[0]
n3

# Selecione os imóveis que tenham pelo menos 4 quartos e aluguel menor que R$ 2.000,00.
selecao = (dados['Quartos'] >= 4) & (dados['Valor'] < 2000)
n4 = dados[selecao].shape[0]
n4

print(f"N⁰ de imóveis classificados com tipo 'Apartamento' -> {n1} ")
print(f"N⁰ de imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila' -> {n2} ")
print(f"N⁰ de imóveis com área entre 60 e 100 metros quadrados, incluindo os limites -> {n3} ")
print(f"N⁰ de imóveis que tenham pelo menos 4 quartos e aluguel menor que R$ 2.000,00 -> {n4} ")

## Exercícios

alunos = pd.DataFrame({'Nome': ['Ary', 'Cátia', 'Denis', 'Beto', 'Bruna', 'Dara', 'Carlos', 'Alice'], 
                        'Sexo': ['M', 'F', 'M', 'M', 'F', 'F', 'M', 'F'], 
                        'Idade': [15, 27, 56, 32, 42, 21, 19, 35], 
                        'Notas': [7.5, 2.5, 5.0, 10, 8.2, 7, 6, 5.6], 
                        'Aprovado': [True, False, False, True, True, True, False, False]}, 
                        columns = ['Nome', 'Idade', 'Sexo', 'Notas', 'Aprovado'])

alunos

selecao = alunos['Aprovado'] == True
alunos[selecao]

selecao = (alunos['Aprovado'] == True) & (alunos['Sexo'] == 'F') 
alunos[selecao]

selecao = (alunos.Idade > 10) & (alunos.Idade < 20) | (alunos.Idade >= 40)
alunos[selecao]
