==> Course 4 - Pandas (Analyzing and Processing Data)/5 - Tratamento de Dados Faltantes.ipynb <==
# Relatório de Análise V

## Tratamento de Dados Faltantes

import pandas as pd

dados = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')

dados.head(10)

dados.isnull()

dados.notnull()

dados.info()

dados[dados['Valor'].isnull()]

A = dados.shape[0]
dados.dropna(subset = ['Valor'], inplace = True)
B = dados.shape[0]

A

B

### Tratamento de Dados Faltantes (continuação)

dados[dados['Condominio'].isnull()].shape[0]

selecao = (dados['Tipo'] == 'Apartamento') & (dados['Condominio'].isnull())

A = dados.shape[0]
dados = dados[~selecao]
B = dados.shape[0]
A - B

dados[dados['Condominio'].isnull()].shape[0]

dados = dados.fillna({'Condominio': 0, 'IPTU': 0})

dados[dados['Condominio'].isnull()].shape[0]

dados[dados['IPTU'].isnull()].shape[0]

dados.info()

dados.to_csv('dados/aluguel_residencial.csv', sep = ';', index = False)

dados

## Tratamento de Dados Faltantes (continuação)

dados[dados['Condominio'].isnull()].shape[0]

 selecao = (dados['Tipo'] == 'Apartamento') & (dados['Condominio'].isnull())

A = dados.shape[0]
dados = dados[~selecao]
B = dados.shape[0]
A - B

dados[dados['Condominio'].isnull()].shape[0]

selecao = (dados['Tipo'] == 'Apartamento') & (dados['Condominio'].isnull())

dados.fillna(0, inplace = True)

dados = dados.fillna({'Condominio': 0, 'IPTU': 0})

dados.info() 

# dados.to_csv('dados/aluguel_residencial.csv', sep = ';', index = False)

### Exercício

imoveis = pd.DataFrame([['Apartamento', None, 970, 68], 
                        ['Apartamento', 2000, 878, 112], 
                        ['Casa', 5000, None, 500], 
                        ['Apartamento', None, 1010, 170], 
                        ['Apartamento', 1500, 850, None], 
                        ['Casa', None, None, None], 
                        ['Apartamento', 2000, 878, None], 
                        ['Apartamento', 1550, None, 228], 
                        ['Apartamento', 2500, 880, 195]], 
                        columns = ['Tipo', 'Valor', 'Condominio', 'IPTU'])

imoveis

imoveis['Valor'].isnull().shape[0]

A = imoveis.shape[0]
imoveis.dropna(subset = ['Valor'], inplace = True)
imoveis

# selecao = (imoveis['Tipo'] == 'Apartamento') & (imoveis['Condominio'].isnull()) 

selecao = (imoveis['Tipo'] == 'Apartamento') & (imoveis['Condominio'].isnull())
imoveis = imoveis[~selecao]
imoveis

imoveis = imoveis.fillna({'Condominio': 0, 'IPTU': 0})
imoveis

imoveis.index = range(imoveis.shape[0])
imoveis

imoveis2 = pd.DataFrame([['Apartamento', None, 970, 68], 
                        ['Apartamento', 2000, 878, 112], 
                        ['Casa', 5000, None, 500], 
                        ['Apartamento', None, 1010, 170], 
                        ['Apartamento', 1500, 850, None], 
                        ['Casa', None, None, None], 
                        ['Apartamento', 2000, 878, None], 
                        ['Apartamento', 1550, None, 228], 
                        ['Apartamento', 2500, 880, 195]], 
                        columns = ['Tipo', 'Valor', 'Condominio', 'IPTU'])

imoveis2

teste = imoveis2['Condominio'].notnull() 

teste

selecao2 = (imoveis2['Tipo'] == 'Apartamento') & (imoveis2['Condominio'].isnull())

imoveis2[~selecao2]
