==> Course 4 - Pandas (Analyzing and Processing Data)/6 - Criando Novas Variáveis.ipynb <==
# Relatório de Análise VI

## Criando Novas Variáveis

import pandas as pd

dados = pd.read_csv('dados/aluguel_residencial.csv', sep = ';')
dados.head(10)

dados['Valor Bruto'] = dados['Valor'] + dados['Condominio'] + dados['IPTU']
dados.head(10)

dados['Valor m2'] = dados['Valor'] / dados['Area']
dados

dados['Valor m2'] = dados['Valor m2'].round(2)

dados

dados['Valor Bruto m2'] = (dados['Valor Bruto'] / dados['Area']).round(2)

dados

list(dados['Tipo'].drop_duplicates())

casa = ['Casa', 'Casa de Condominio', 'Casa de Vila']

dados['Tipo Agregado'] = dados['Tipo'].apply(lambda x: 'Casa' if x in casa else 'Apartamento')
dados.head(10)

### Exercício

alunos = pd.DataFrame({'Nome': ['Ary', 'Cátia', 'Denis', 'Beto', 'Bruna', 'Dara', 'Carlos', 'Alice'], 
                        'Sexo': ['M', 'F', 'M', 'M', 'F', 'F', 'M', 'F'], 
                        'Idade': [15, 27, 56, 32, 42, 21, 19, 35], 
                        'Notas': [7.5, 2.5, 5.0, 10, 8.2, 7, 6, 5.6]}, 
                        columns = ['Nome', 'Idade', 'Sexo', 'Notas'])

alunos

alunos['Notas-Média(Notas)'] = alunos['Notas'].apply(lambda x: x - alunos['Notas'].mean())
alunos

alunos['Faixa Etária'] = alunos['Idade'].apply(lambda x: 'Menor que 20 anos' if x < 20 else 
                                               ('Entre 20 e 40 anos' if (x >= 20 and x <= 40) else 'Maior que 40 anos'))

alunos

alunos['Notas-Média(Notas)'] = alunos.Notas - alunos.Notas.mean()
alunos

## Excluindo Variáveis

dados_aux = pd.DataFrame(dados[['Tipo Agregado', 'Valor m2', 'Valor Bruto', 'Valor Bruto m2']])
dados_aux.head(10)

del dados_aux['Valor Bruto']
dados_aux.head(10)

dados_aux.pop('Valor Bruto m2')

dados_aux

# dados.drop(['Valor Bruto', 'Valor Bruto m2'], axis = 1, inplace = True)

dados

dados.to_csv('dados/aluguel_residencial.csv', sep = ';', index = False)

dados
